+++ src/main.cpp
@@ -155,7 +155,7 @@ CircleButtonSprite* createCheatIndicator(bool isHacking) {
     return cheatIndicator;
 }
 
-#ifdef GEODE_IS_WINDOWS
+#ifndef GEODE_IS_MACOS
 void drawPlayerHitboxes(CCDrawNode* drawNode, PlayerObject* player) {
     auto innerRect = player->getObjectRect(0.25f, 0.25f);
     drawNode->drawRect(
@@ -462,7 +462,7 @@ class $modify(PlayLayer) {
     }
 
     // Show Hitboxes
-#ifdef GEODE_IS_WINDOWS
+#ifndef GEODE_IS_MACOS
     void updateVisibility(float p0) {
         PlayLayer::updateVisibility(p0);
         if (Hacks::isHackEnabled("Show Hitboxes") || (Hacks::isHackEnabled("Show Hitboxes on Death") && m_player1->m_isDead)) {
@@ -477,7 +477,7 @@ class $modify(PlayLayer) {
 
 };
 
-#ifdef GEODE_IS_WINDOWS // Show Hitboxes
+#ifndef GEODE_IS_MACOS // Show Hitboxes
 
 #include <Geode/modify/LevelEditorLayer.hpp>