+++ changelog.md
@@ -8,8 +8,8 @@
     - Added Suicide.
     - Added Disable Camera Effects
     - Added Turkish Translations.
-    - Allowed Mac OS to view Graphic Options.
-    - Fixed FPS Bypass for Mac OS.
+    - Fixed "Solid Wave Trail" not working for Mac OS.
+    - Fixed "No Transition" not working for Mac OS.
     - Fixed bug with Jump Hack not working when upside down.
     - Fixed issue with "Show Graphic Options" not working on ImGui.
     - Fixed issue with Accurate Percentage not working sometimes in Test Mode.